Proposal to extend and fix BusinessFactory
See Issue #176
This is a proposal.

Note that API is changing a bit since session was not a fixed argument
while it should be. Since we are not in Java 1.8, default Interface
implementation is not possible, therefore introducing API backward
compatibility.

The idea is to propose another "simple" way to track transfers, from
connection, authentication then up to normal steps of a transfers.
Additionnal methods are provided.

An example is provided, which only logs the activities in Warn mode.
To test the behaviour, simply add the definition as follow in your
configuration file:

<server>
...

<businessfactory>org.waarp.openr66.context.task.test.TestBusinessFactory</businessfactory>
</server>