Fix compilation with Java 1.8
Java 1.8 changed the signature of ConcurrentHashMap.keySet() from

```
Set<K> keySet()
```

to

```
ConcurrentHashMap.KeySetView keySet()
```

When compiled with java 1.8, it breaks Maven's animal-sniffer-plugin,
even cross-compiling with -source 1.7 -target 1.7.

In this case, we were interested in having a Set<String> value,
so casting the ConcurentHashMap object to the Map interface before
calling keyset() solves the compatibility with java 1.7 issues.