Removes the error when changing from ERROR to DATAS
The test to check whether data packets must be sent is done in the loop
condition (running.get()). However, if the stop order is given while the future
is waiting (which effectively pause the loop), a new packet was sent, even if it
should not. This caused an error when the internal state machine tried to go
from ERROR to DATAS.

Checking the value of running _after_ the promise has returned prevents this
packet to be sent and removes the error.